* #13: Auth files code and structure enhancements along with layout normalization.     - tsconfig.app.json: solved error regarding ignoreDeprecations version.     - index.css: defined custom utility tailwind classes for consistent styling across the app.         - Some of them are: input, form, flex-align, flex-center, flex-between, label, etc...         - They will be used to normalize form styles across auth and other components.     - Footer.tsx: Changed my github account url to my portfolio url.     - AuthLayout.tsx: Added base layout styles to be applied across all auth views.     - SetPasswordForm.tsx: Renamed to PasswordForm.tsx and updated to use own custom utility classes for styling.
