+++ src/layouts/AuthLayout.tsx
@@ -7,12 +7,12 @@ export default function AuthLayout() {
     return (
         <>
             <div className="bg-orange-500 min-h-screen flex items-center justify-center px-6">
-                <div className="py-10 lg:py-20 mx-auto w-full max-w-md md:max-w-2xl 2xl:max-w-4xl">
+                <div className="py-10 lg:py-20 mx-auto w-full md:max-w-2xl 2xl:max-w-4xl">
                     <div className="flex justify-center">
                         <Logo />
                     </div>
 
-                    <div className="">
+                    <div className="mx-auto max-w-2xl lg:px-0">
                         <Outlet />
                     </div>
                 </div>
