+++ tsconfig.app.json
@@ -1,6 +1,6 @@
 {
 	"compilerOptions": {
-		"ignoreDeprecations": "5.0",
+		"ignoreDeprecations": "6.0",
 		"tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
 		"target": "ES2020",
 		"useDefineForClassFields": true,
